lab1/requirements.txt
aws-cdk-lib==2.164.1
constructs>=10.0.0
cdk-ecr-deployment==3.0.128
python-dotenv==1.0.1
boto3==1.18.63